requirements.txt
django>=5.0
djangorestframework
celery
redis
requests
groq
python-dotenv 